Improved multithreading and .obj file parsing

+++ RT/src/scene/PreviewScene.java
@@ -7,17 +7,10 @@ package scene;
 import geometry.Pt;
 import geometry.Vec;
 
-import java.io.IOException;
-import java.util.List;
-
 import javax.vecmath.AxisAngle4d;
-import javax.vecmath.Point3d;
 import javax.vecmath.Vector3d;
 
-import objects.SceneObject;
 import objects.Sphere;
-import objects.TriangleMesh;
-import parser.ObjectParser;
 import raytracer.Camera;
 
 /**
@@ -30,7 +23,7 @@ public class PreviewScene extends Scene
 		super();
 
 		camera = new Camera(new Pt(0, 0, -5), new Pt(0, 0, 300), new Vec(0, 1, 0),
-				(double) (Math.PI / 4));
+				Math.PI / 4);
 		PointLight demoLight = new PointLight();
 		PointLight demoLight2 = new PointLight();
 		demoLight.setPosition(new Vector3d(100, 100, 50));
